Rename month assembly package to closedmonth
